Use the canonical uint32 as integer type in epochid package
We need some conversion to int64 in the database code, but can get rid of the
error handling here.

+++ rolling-shutter/decryptor/inputhandling.go
@@ -105,7 +105,7 @@ func handleSignatureInput(
 	}
 
 	activationBlockNumber := epochid.BlockNumber(signature.epochID)
-	decryptorSet, err := db.GetDecryptorSet(ctx, activationBlockNumber)
+	decryptorSet, err := db.GetDecryptorSet(ctx, int64(activationBlockNumber))
 	if err != nil {
 		return nil, errors.Wrapf(err, "failed to query decryptor set from db")
 	}
